une composante vue WineListItem, qui est appelé de la composante vue WineList
